// src/app/core/models/work-order.model.ts
export type WorkOrderStatus = |'open'|'in-progress'|'complete'|'blocked';
export interface WorkOrderDocument {
    docId: string;
    docType: 'workOrder';
    data: {
        name: string;
        workCenterId: string;
        status: WorkOrderStatus;
        startDate:  string;
        endDate: string;
    };
}